+++ docs/changelog.md
@@ -251,6 +251,7 @@ Many feature gates are removed in order to make most features make always availa
 
 #### iter
 - new traits: `IteratorLending`, `IteratorLendingDoubleEnded`, `IteratorLendingExactSize`, `IteratorLendingPeek`, `IteratorLendingPeekDoubleEnded`.
+- new macro: `iter_strided!`.
 - new types: `StridedIter`, `StridedIterMut`.
 
 ### codec